chore: moved createAuth into src/client
