+++ src/client/index.ts
@@ -4,6 +4,28 @@ import {
   postSignout,
 } from './generatedRoutes'
 import {getAPIBasepath} from 'src/utils/basepath'
+import {Authorization} from 'src/types'
+import {postAuthorization} from 'src/client'
+
+
+export const createAuthorization = async (
+  authorization
+): Promise<Authorization> => {
+  try {
+    const resp = await postAuthorization({
+      data: authorization,
+    })
+
+    if (resp.status !== 201) {
+      throw new Error(resp.data.message)
+    }
+    
+    return resp.data
+  } catch (error) {
+    console.error(error)
+    throw error
+  }
+}
 
 setRequestHandler((url: string, query: string, init: RequestInit) => {
   return {
@@ -23,5 +45,7 @@ setResponseHandler((status, headers, data) => {
 
   return {status, headers, data}
 })
+
+
 
 export * from './generatedRoutes'
